+++ try-clients/tests/github.ts
@@ -1,4 +1,5 @@
 import { createClient, everything } from "../github/generated";
+import { AxiosInstance } from "axios";
 
 describe("github", () => {
   const token = process.env.GITHUB_TOKEN;
@@ -11,6 +12,18 @@ describe("github", () => {
       Authorization: "Bearer " + token,
     },
   });
+  // example of assigning a custom fetcher and interceptor ( Axios Based )
+  const axiosClient = client.fetcherInstance as AxiosInstance;
+  axiosClient.interceptors.request.use(
+    (config) => {
+      console.log(config);
+      return config;
+    },
+    (error) => {
+      console.log(error);
+      return Promise.reject(error);
+    }
+  );
   it("simple normal syntax", async () => {
     const res1 = await client.query({
       repository: [